% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.list.R
\name{as.list,VennPlotData-method}
\alias{as.list,VennPlotData-method}
\title{Transform a VennPlotData object to a generic list}
\usage{
\S4method{as.list}{VennPlotData}(x)
}
\arguments{
\item{x}{a VennPlotData class object.}
}
\value{
a list
}
\description{
So that users can use this object without shapeMageR package.
}
\examples{
 f4e = VennPlotData(
  shapeId = "401f",
  type = "ellipse",
  setEdge = fancy_4d_ellipse(),
  setLabel =  fancy_4d_ellipse_label()
  )
 list = as.list(f4e)
 names(f4e)
}
