% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{discern}
\alias{discern}
\alias{discern,Polygon-method}
\title{Discern the 'true' unique region of \code{Polygon} object}
\usage{
discern(polygon, slice1, slice2)

\S4method{discern}{Polygon}(polygon, slice1, slice2 = "all")
}
\arguments{
\item{polygon}{Polygon object}

\item{slice1}{first slice of Polygon object}

\item{slice2}{second slice of Polygon object, default is all except the first slice}
}
\value{
a Polygon object
}
\description{
Discern the 'true' unique region of \code{Polygon} object
}
\examples{
# don't run
# discern(polygon, slice1 = 1)
}
