% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{discern_overlap}
\alias{discern_overlap}
\title{Calculate region of polygons}
\usage{
discern_overlap(venn, slice = "all")

\S4method{discern_overlap}{Polygon}(venn, slice = "all")
}
\arguments{
\item{slice}{index of Venn members, default is "all"}

\item{polygon}{a Polygon object}
}
\value{
region items
}
\description{
Calculate region of polygons

calculate the unique region defined by \code{Venn} object and the parameter \code{slice}
}
