% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{overlap}
\alias{overlap}
\alias{overlap,Polygon-method}
\title{calculate the overlapping region of \code{Polygon} object}
\usage{
overlap(polygon, slice = "all")

\S4method{overlap}{Polygon}(polygon, slice = "all")
}
\arguments{
\item{polygon}{a Polygon object}

\item{slice}{index of Venn members, default is "all"}
}
\value{
a Polygon object
}
\description{
calculate the overlapping region of \code{Polygon} object
}
\examples{
  library("sf")

 # a rectangle
 v1 = rbind(c(0,0),c(0,1),c(1,1),c(1,0),c(0,0))
 rect = st_polygon(list(v1))

 # a triangle
 v2 = rbind(c(0.5,0.5),c(1.5,1.5),c(1,0),c(0.5,0.5))
 tri = st_polygon(list(v2))


 # a circle

 # overlapping region
 # polygons = ....
 # overlap = overlap(rect, tri, circle)
 # plot(overlap)
}
