% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_basic.R
\name{polygon}
\alias{polygon}
\alias{triangle}
\alias{rectangle}
\title{generating a polygon such as triangle and rectangle by given vertex}
\usage{
triangle(xy = c(0, 0, 1, 0, 0, 1))

rectangle(xy = c(0, 0, 0, 1, 1, 1, 1, 0))
}
\arguments{
\item{xy}{coordinates of the vertex defining a polygon}
}
\value{
a matrix with xy coordinates
}
\description{
generating a polygon such as triangle and rectangle by given vertex
}
\examples{
# triangle coordinates
triangle()

# plot a new triangle
t = triangle()
plot(t, type = "l")

# plot a new rectangle
r = rectangle()
plot(r, type = "l")
}
