% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{slice_idx}
\alias{slice_idx}
\title{check and format slice name}
\usage{
slice_idx(polygon, slice)
}
\arguments{
\item{polygon}{a Polygon object}

\item{slice}{a numeric or character vector}
}
\value{
the index of polygon \link{numeric vector} or "all"
}
\description{
check and format slice name
}
