% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_basic.R
\name{triangle}
\alias{triangle}
\title{generating a triangle by three points}
\usage{
triangle(xy = c(0, 0, 1, 0, 0, 1))
}
\arguments{
\item{xy}{coordinates of the three points defining a triangle}
}
\value{
a matrix with xy coordinates
}
\description{
generating a triangle by three points
}
\examples{
# triangle coordinates
library(sf)
triangle()

# plot a new triangle
triangle() \%>\% st_linestring() \%>\% plot()
}
