% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclass.R
\name{unclass}
\alias{unclass}
\title{Remove shape class}
\usage{
unclass(obj)
}
\arguments{
\item{obj}{a VennPlotData object}
}
\value{
a list
}
\description{
Unlink shapes from the 'shapeMageR' package.
}
\examples{
# 4d ellipse
f4e = VennPlotData(
  shapeId = "401f",
  type = "ellipse",
  setEdge = fancy_4d_ellipse(),  # how to store object in data.frame
  setLabel =  fancy_4d_ellipse_label())
l = unclass(f4e)
l
}
