% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{Venn-method}
\alias{Venn-method}
\alias{Venn,ANY-method}
\title{Venn class object constructor}
\usage{
Venn(sets, names)

\S4method{Venn}{ANY}(sets, names)
}
\arguments{
\item{sets}{a list of sets}

\item{names}{names of sets}
}
\description{
Venn class object constructor
}
