% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\docType{methods}
\name{VennPlotData}
\alias{VennPlotData}
\alias{VennPlotData,ANY-method}
\title{VennPlotData constructor}
\usage{
VennPlotData(shapeId, type, setEdge, setLabel)

\S4method{VennPlotData}{ANY}(shapeId, type, setEdge, setLabel)
}
\arguments{
\item{shapeId}{shape id}

\item{type}{type of shape, can be one of ellipse, circle, triangle, or polygon}

\item{setEdge}{a list of coordinates matrix defining Venn set edges}

\item{setLabel}{a list of coordinates matrix defining Venn set labels}
}
\value{
a S4 class VennPlotData object
}
\description{
Region Edge and Label will be caculated automatically with functions from \code{sf} package.
}
\examples{
# construct a VennPlotData
venn = VennPlotData(shapeId = "1",
                    type = "e",
                    setEdge = fancy_4d_ellipse(),
                    setLabel = fancy_4d_ellipse_label())
}
