% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{discern_overlap}
\alias{discern_overlap}
\alias{discern_overlap,Polygon-method}
\title{Calculate the 'unique' region defined by \code{Venn} object and the parameter \code{slice}}
\usage{
discern_overlap(polygon, slice = "all")

\S4method{discern_overlap}{Polygon}(polygon, slice = "all")
}
\arguments{
\item{polygon}{a Polygon object}

\item{slice}{index of Venn members, default is "all"}
}
\value{
region items
}
\description{
Calculate the 'unique' region defined by \code{Venn} object and the parameter \code{slice}
}
\examples{
# discern_overlap(polygon)
}
