% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unclass.R
\name{unclass}
\alias{unclass}
\title{Remove shape class}
\usage{
unclass(obj)
}
\arguments{
\item{obj}{a VennShape object}
}
\value{
a list, the items are data.frame for plot
}
\description{
Unlink shapes from the 'shapeMageR' package.
}
\examples{
# 4d ellipse
f4e = VennShape("401f", "ellipse", fancy_4d_ellipse(), fancy_4d_ellipse_label())
l = unclass(f4e)
l
}
